+++ omnisafe/algorithms/offline/base.py
@@ -76,7 +76,7 @@ class BaseOffline(BaseAlgo):
         self._logger.register_key('Train/Epoch')
         self._logger.register_key('TotalSteps')
 
-    def learn(self) -> tuple[float, float, int]:
+    def learn(self) -> tuple[float, float, float]:
         """Learn the policy."""
         self._logger.log('Start training ...')
 
@@ -112,7 +112,7 @@ class BaseOffline(BaseAlgo):
 
         ep_ret = self._logger.get_stats('Metrics/EpRet')[0]
         ep_cost = self._logger.get_stats('Metrics/EpCost')[0]
-        ep_len = int(self._logger.get_stats('Metrics/EpLen')[0])
+        ep_len = self._logger.get_stats('Metrics/EpLen')[0]
         self._logger.close()
 
         return ep_ret, ep_cost, ep_len
