+++ omnisafe/algorithms/on_policy/base/policy_gradient.py
@@ -222,7 +222,7 @@ class PolicyGradient(BaseAlgo):
         self._logger.register_key('Time/Epoch')
         self._logger.register_key('Time/FPS')
 
-    def learn(self) -> tuple[float, float, int]:
+    def learn(self) -> tuple[float, float, float]:
         """This is main function for algorithm update.
 
         It is divided into the following steps:
@@ -282,7 +282,7 @@ class PolicyGradient(BaseAlgo):
 
         ep_ret = self._logger.get_stats('Metrics/EpRet')[0]
         ep_cost = self._logger.get_stats('Metrics/EpCost')[0]
-        ep_len = int(self._logger.get_stats('Metrics/EpLen')[0])
+        ep_len = self._logger.get_stats('Metrics/EpLen')[0]
         self._logger.close()
 
         return ep_ret, ep_cost, ep_len
